+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # NHS digital service manual Changelog
 
+## 2.2.0 - 13 January 2020
+
+:new: **New features**
+
+- Display the parent page in search results
+
+:wrench: **Fixes**
+
+- Update package dependencies to latest versions
+
 ## 2.1.0 - 6 January 2020
 
 :new: **New features**